Move templates outside the logic package, breaking cyclical import
